test: split manage tasks support mocks and unit coverage
